Fix share permission l18n & icons
